Fix narrowing conversion warnings on GCC 10.1
Using an unsigned value in a switch with signed case values produced the
error.

+++ maxutils/maxbase/src/worker.cc
@@ -622,7 +622,7 @@ void Worker::shutdown()
  */
 void Worker::handle_message(MessageQueue& queue, const MessageQueue::Message& msg)
 {
-    switch (msg.id())
+    switch ((int)msg.id())
     {
     case MXB_WORKER_MSG_SHUTDOWN:
         {
